Add files via upload
The Greco-Roman culture was able to classify a great variety of gems. Authors such as Theophrastus, Plutarch and Pliny the Elder dealt with the subject. To Pliny, in particular, belongs the most extensive work on the knowledge of precious stones in antiquity (Book 37 of his Naturalis Historia). Pliny's book served in turn as a basis for the writing of other corpora on gems, being the Isidore of Seville book’s the best example. To know which the most appreciated gems in ancient Rome were, it’s essential to use the book 37 of Pliny the Elder, who catalogued the most prestigious gems, their origin and their qualities. The corpus collects a total of 240 different variants of gems, of which, in 93 cases, its place of origin is known. Of the gems from which its origin is known, special mention deserves the precious stones from North Africa and the territories to the east of the Empire. In this line, there was a clear predominance of gems from India, Persia , Egypt , Arabia  and Africa , being the most desired by the Roman elites. Besides these regions, they were also significant in the exploitation of these mining resources, Greece , the islands of the eastern Mediterranean  and the territories of Asia Minor (now Turkey)